move attribute parsing into util class
