+++ otel/tests/phpt/trace/span/span-set-attributes.phpt
